requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cmake==3.24.1.1
contourpy==1.0.5
cycler==0.11.0
dlib==19.24.0
docopt==0.6.2
face-recognition==1.3.0
face-recognition-models==0.3.0
ffmpeg-python==0.2.0
fonttools==4.37.4
future==0.18.2
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.1
numpy==1.23.3
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
tqdm==4.64.1
urllib3==1.26.12
yarg==0.1.9
